Restructuring the new folder... fixing imports so that tests, notebooks all work. Closes #7 and closes #8.
