+++ hydradx_update/README.md
@@ -1,0 +1,3 @@
+# Modeling Omnipool in cadCAD
+- TestSwap.ipynb demonstrates how cadCAD can be run
+- The "spec" folder has the mathematical spec of the implemented AMM
