--- hydradx_update/TestSwap.py
@@ -1,70 +1,0 @@
-import pandas as pd
-
-from hydradx_update.model import init_utils
-from hydradx_update.model import processing
-# Experiments
-from hydradx_update.model import run
-from hydradx_update.model.plot_utils import *
-
-########## AGENT CONFIGURATION ##########
-# key -> token name, value -> token amount owned by agent
-# note that token name of 'omniABC' is used for omnipool LP shares of token 'ABC'
-# omniHDXABC is HDX shares dedicated to pool of token ABC
-
-trader = {'HDX': 1000000, 'R1': 1000000, 'R2': 1000000}
-
-# key -> agent_id, value -> agent dict
-agent_d = {'Trader': trader}
-
-########## ACTION CONFIGURATION ##########
-
-action_dict = {
-    'sell_r2_for_r1': {'token_buy': 'R1', 'token_sell': 'R2', 'amount_sell': 5000, 'action_id': 'Trade',
-                       'agent_id': 'Trader'},
-    'sell_r1_for_r2': {'token_sell': 'R1', 'token_buy': 'R2', 'amount_sell': 1000, 'action_id': 'Trade',
-                       'agent_id': 'Trader'}
-}
-
-# list of (action, number of repetitions of action), timesteps = sum of repititions of all actions
-trade_count = 1000
-action_ls = [('trade', trade_count)]
-
-# maps action_id to action dict, with some probability to enable randomness
-prob_dict = {
-    'trade': {'sell_r2_for_r1': 0.5,
-              'sell_r1_for_r2': 0.5}
-}
-
-########## CFMM INITIALIZATION ##########
-
-initial_values = {
-    'token_list': ['R1', 'R2'],
-    'R': [500000, 1500000],
-    'P': [2, 2 / 3]
-}
-
-############################################ SETUP ##########################################################
-
-config_params = {
-    'cfmm_type': "",
-    'initial_values': initial_values,
-    'agent_d': agent_d,
-    'action_ls': action_ls,
-    'prob_dict': prob_dict,
-    'action_dict': action_dict,
-}
-
-config_dict, state = init_utils.get_configuration(config_params)
-
-pd.options.mode.chained_assignment = None  # default='warn'
-pd.options.display.float_format = '{:.2f}'.format
-
-run.config(config_dict, state)
-events = run.run()
-
-rdf, agent_df = processing.postprocessing(events)
-
-print(agent_df.tail(20))
-
-var_list = ['R', 'Q']
-plot_vars(rdf, var_list)
