+++ hydradx_update/requirements.txt
@@ -1,0 +1,4 @@
+cadcad
+hypothesis
+matplotlib
+pandas
